+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/cryptostore/db/RealmCryptoStoreMigration.kt
@@ -18,14 +18,17 @@ package org.matrix.androidsdk.data.cryptostore.db
 
 import io.realm.DynamicRealm
 import io.realm.RealmMigration
+import org.matrix.androidsdk.data.cryptostore.db.model.IncomingRoomKeyRequestEntityFields
+import org.matrix.androidsdk.data.cryptostore.db.model.KeysBackupDataEntityFields
 import org.matrix.androidsdk.data.cryptostore.db.model.OlmSessionEntityFields
+import org.matrix.androidsdk.data.cryptostore.db.model.OutgoingRoomKeyRequestEntityFields
 import org.matrix.androidsdk.util.Log
 
 internal object RealmCryptoStoreMigration : RealmMigration {
 
     private const val LOG_TAG = "RealmCryptoStoreMigration"
 
-    const val CRYPTO_STORE_SCHEMA_VERSION = 1L
+    const val CRYPTO_STORE_SCHEMA_VERSION = 2L
 
     override fun migrate(realm: DynamicRealm, oldVersion: Long, newVersion: Long) {
         Log.d(LOG_TAG, "Migrating Realm Crypto from $oldVersion to $newVersion")
@@ -39,6 +42,68 @@ internal object RealmCryptoStoreMigration : RealmMigration {
                     ?.transform {
                         it.setLong(OlmSessionEntityFields.LAST_RECEIVED_MESSAGE_TS, 0)
                     }
+        }
+
+        if (oldVersion <= 1) {
+            Log.d(LOG_TAG, "Step 1 -> 2")
+            Log.d(LOG_TAG, "Update IncomingRoomKeyRequestEntity format: requestBodyString field is exploded into several fields")
+
+            realm.schema.get("IncomingRoomKeyRequestEntity")
+                    ?.addField(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_ALGORITHM, String::class.java)
+                    ?.addField(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_ROOM_ID, String::class.java)
+                    ?.addField(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_SENDER_KEY, String::class.java)
+                    ?.addField(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_SESSION_ID, String::class.java)
+                    ?.transform { dynamicObject ->
+                        val requestBodyString = dynamicObject.getString("requestBodyString")
+                        try {
+                            // It was a map before
+                            val map: Map<String, String>? = deserializeFromRealm(requestBodyString)
+
+                            map?.let {
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_ALGORITHM, it["algorithm"])
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_ROOM_ID, it["room_id"])
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_SENDER_KEY, it["sender_key"])
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_SESSION_ID, it["session_id"])
+                            }
+                        } catch (e: Exception) {
+                            Log.d(LOG_TAG, "Error", e)
+                        }
+                    }
+                    ?.removeField("requestBodyString")
+
+            Log.d(LOG_TAG, "Update IncomingRoomKeyRequestEntity format: requestBodyString field is exploded into several fields")
+
+            realm.schema.get("OutgoingRoomKeyRequestEntity")
+                    ?.addField(OutgoingRoomKeyRequestEntityFields.REQUEST_BODY_ALGORITHM, String::class.java)
+                    ?.addField(OutgoingRoomKeyRequestEntityFields.REQUEST_BODY_ROOM_ID, String::class.java)
+                    ?.addField(OutgoingRoomKeyRequestEntityFields.REQUEST_BODY_SENDER_KEY, String::class.java)
+                    ?.addField(OutgoingRoomKeyRequestEntityFields.REQUEST_BODY_SESSION_ID, String::class.java)
+                    ?.transform { dynamicObject ->
+                        val requestBodyString = dynamicObject.getString("requestBodyString")
+                        try {
+                            // It was a map before
+                            val map: Map<String, String>? = deserializeFromRealm(requestBodyString)
+
+                            map?.let {
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_ALGORITHM, it["algorithm"])
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_ROOM_ID, it["room_id"])
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_SENDER_KEY, it["sender_key"])
+                                dynamicObject.setString(IncomingRoomKeyRequestEntityFields.REQUEST_BODY_SESSION_ID, it["session_id"])
+                            }
+                        } catch (e: Exception) {
+                            Log.d(LOG_TAG, "Error", e)
+                        }
+                    }
+                    ?.removeField("requestBodyString")
+
+            Log.d(LOG_TAG, "Create KeysBackupDataEntity")
+
+            realm.schema.create("KeysBackupDataEntity")
+                    .addField(KeysBackupDataEntityFields.PRIMARY_KEY, Integer::class.java)
+                    .addPrimaryKey(KeysBackupDataEntityFields.PRIMARY_KEY)
+                    .setRequired(KeysBackupDataEntityFields.PRIMARY_KEY, true)
+                    .addField(KeysBackupDataEntityFields.BACKUP_LAST_SERVER_HASH, String::class.java)
+                    .addField(KeysBackupDataEntityFields.BACKUP_LAST_SERVER_NUMBER_OF_KEYS, Integer::class.java)
         }
     }
 }